(maint) Build either rpm or swix packages when building for EOS
